src/test/javascript/spec/app/entities/gap-analysis/gap-analysis.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { GapAnalysisService } from 'app/entities/gap-analysis/gap-analysis.service';
import { IGapAnalysis, GapAnalysis } from 'app/shared/model/gap-analysis.model';

describe('Service Tests', () => {
  describe('GapAnalysis Service', () => {
    let injector: TestBed;
    let service: GapAnalysisService;
    let httpMock: HttpTestingController;
    let elemDefault: IGapAnalysis;
    let expectedResult: IGapAnalysis | IGapAnalysis[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(GapAnalysisService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new GapAnalysis(0, 'AAAAAAA', 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a GapAnalysis', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new GapAnalysis()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a GapAnalysis', () => {
        const returnedFromService = Object.assign(
          {
            codiceEsco: 'BBBBBB',
            nome: 'BBBBBB',
            frequenza: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of GapAnalysis', () => {
        const returnedFromService = Object.assign(
          {
            codiceEsco: 'BBBBBB',
            nome: 'BBBBBB',
            frequenza: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a GapAnalysis', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
